Fetch watchlist data in home component
Instead of fetching the watchlist data in the watchlist component, we
fetch the data when the home component is loaded. This way the browse
component will have no problems depending on the data it needs.

The last implementation where the data was fetch on the watchlist
component, the browse component gets nothing since the watchlist is
not mounted.